c2-ws/requirements.txt
websockets==12.0
redis==5.0.1
aioredis==2.0.1
PyYAML==6.0